o Applying patch provided by John Thorhauer:
I have a patch for betwixt.  The problem that I am having is with the
DefaultPluralStemmer.  It does not deal real well with plurals that end
in 'es'.  I have included a patch for both the test case and the patch
for the DefaultPluralStemmer.  If you run the test without applying the
fix you will see the problem.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/betwixt/trunk@128959 13f79535-47bb-0310-9956-ffa450edef68