Added provisional support to allow bean types to be customized by the presence of a 'className' attribute.
Its fairly simple and only works for parsing of beans right now. We need to add this customization feature to the ElementDescriptors so that this can be enabled as required and 
then used to auto-default an AttributeDescriptor with a ClassNameExpression so that the output of the bean will also convey the className attribute again.

Also the attribute used to describe the class name should be configurable.

Maybe we should default it to 'class' since its rare to ever get a bean with a 'class' setter property?


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/betwixt/trunk@129106 13f79535-47bb-0310-9956-ffa450edef68